undo osg-release rpm gpg key extraction
svn merge -c-21037 .

Turns out this didn't work so well on the repo* servers, which run el5,
because apparently the el6 rpms use a newer format of cpio which is *not*
supported by the version of cpio available on el5.  Thus, only the el5 gpg
keys get extracted.

This is all kind of overkill anyway since we don't ever seem to update the
RPM-GPG-KEY-OSG, so there's not really a big need to try extracting it every
time there is a new osg-release.  It should suffice to just copy the new one
to the repo server if/when we ever update it.

https://ticket.grid.iu.edu/25277



git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-next-tools@21059 4e558342-562e-0410-864c-e07659590f8c